app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fh-plotly>=0.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-fasthtml>=0.12.12",
]
